refactor(docker/containers): gpu fieldset
