+++ app/react/docker/containers/CreateView/ResourcesTab/GpuFieldset/index.ts
@@ -1,0 +1,14 @@
+import { validation } from './validation';
+import { toRequest } from './toRequest';
+import { toViewModel, getDefaultViewModel } from './toViewModel';
+
+export { GpuFieldset } from './GpuFieldset';
+
+export type { Values as GpuFieldsetValues } from './types';
+
+export const gpuFieldsetUtils = {
+  toRequest,
+  toViewModel,
+  validation,
+  getDefaultViewModel,
+};
