+++ app/react/docker/containers/CreateView/ResourcesTab/GpuFieldset/types.ts
@@ -1,0 +1,6 @@
+export interface Values {
+  enabled: boolean;
+  useSpecific: boolean;
+  selectedGPUs: string[];
+  capabilities: string[];
+}
