+++ app/react/docker/containers/CreateView/ResourcesTab/GpuFieldset/validation.ts
@@ -1,0 +1,14 @@
+import { SchemaOf, object, array, string, bool } from 'yup';
+
+import { Values } from './types';
+
+export function validation(): SchemaOf<Values> {
+  return object({
+    capabilities: array()
+      .of(string().default(''))
+      .default(['compute', 'utility']),
+    enabled: bool().default(false),
+    selectedGPUs: array().of(string()).default(['all']),
+    useSpecific: bool().default(false),
+  });
+}
